requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.3.0
astunparse==1.6.3
asyncmy==0.2.9
attrs==23.2.0
base58==2.1.1
black==24.3.0
cachetools==4.2.4
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
databases==0.9.0
dateparser==1.2.0
flatbuffers==24.3.7
frozenlist==1.4.1
gast==0.5.4
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h5py==3.10.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
iniconfig==2.0.0
joblib==1.3.2
jsonalias==0.1.1
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.10.0
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.4.0
protobuf==4.25.3
pycoingecko==3.1.0
pycryptodome==3.20.0
Pygments==2.17.2
pytest==8.1.1
python-binance==1.0.19
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
solana==0.32.0
solders==0.20.0
SQLAlchemy==2.0.28
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
types-cachetools==4.2.10
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
websockets==11.0.3
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
yarl==1.9.4
